Add a real test case for bitfield.
This currently fails because we're not linking the test with the objects we want
to test. I tried adding the $(OBJS) variable to the build line, but because
those objects are all compiled with pic32-gcc it's not going to work.

I think the solution is to either:

Manually list the objects we want to recompile and link with (boo)

Compile our non-Arduino code into a static library that we link with.