Attempt to buffer 4 packets of an OUT message.
A simpler approach might be to just blindly copy the entire packet into a 4
packet buffer and then hand the 4 packet buffer off to cJSON. If it can parse
out a JSON message, great - we're done. If not, keep adding to the buffer until
we get to the 5th packet, at which point we have to dump the buffer and start
fresh.