Don't use slow Serial writes when handling EP0 requests.
This seems to alleviate USB connectivity problems with Android. Serial is pretty
slow, so it may be something about not returning a response to an unrecognized
EP0 request in a certain amount of time.

The behavior before this commit was that the CAN translator would trigger the
attached/detached events in Android about 5-15 times just fine. On the last
time, it would take ~20 seconds for the ATTACHED event to happen, then DETACHED
would be triggered as soon as you unplug. After that, no more events are
triggered, ever. Even if you power cycle the CAN translator and restart the
OpenXC service, nothing is even sent. You have to reboot the entire tablet.

Without this call to Serial.println, you can unplug and replug the device 40+
times with no issue. I didn't go higher than that.