Fix a data truncation bug with bit field manipulation.
Also, make sure it never happens again! We were compiling the tests in a 64-bit
environment, where an unsigned long is 64 bits. Everything was working great,
tests passing. On the PIC32 however, I couldn't figure out why one behavior was
incorrect - of course the PIC32 has 32 bit unsigned longs, so the setBitField
function was buggy.

I fixed this specific bug and switched the tests to compile specifically for a
32-bit architecture on the development machine.