Add a lookup function to find writable signals.
This is useful if you use the same generic name both a read-only command and a
write. If the write is on a different message, you want to specifically find
that one and not the read-only message. This minimizes the number of generic
names you must keep track of (e.g. you can share a class in Android for both
reading and writing a measurement, even if it's in separate CAN messages in
either direction).