Configure PIC32 CAN filter to filter on IDs.
This fixes a regression where we weren't actually filtering any IDs. Previously
we were dynamically initializing the mask values, but we didn't have a smart way
to make the masks so they were just 0xfff. I removed this since it didn't seem
to do anything, but I didn't notice that it caused us to receive every CAN
message - not just the one's we are interested in.