Add an extra defense against sending over USB when detached.
If you have USB attached but are not actively reading the stream, it will
slow down the other output interfaces (e.g. UART) because we have to try sending
and wait a little while to see if the transfer is happening. If you detach USB,
we can detect that now (if you have the jumper going to analog input 0), so the
speed will go back up to 100%.