Split the representation of CAN messages into 2 variants to save RAM.
We never fill in most of the metadat in the former CanMessage struct
when we're just using it to hold an ID and data field for a buffered
message, so instead we split it into two variants - the compact
CanMessage for these buffers, and the larger CanMessageDefinition for
storing the metadata.

This saves RAM in both the CanBus.sendQueue and CanBus.receiveQueue, and
it makes the definition of CAN messages smaller by 4 bytes (no 'data'
field) - further savings adding to #162.