Support extended CAN frames (29-bit IDs).
The CanMessage, CanMessageDefinition and AcceptanceFilterListEntry
structs now have a CanMessageFormat attribute, which indicates if the ID
should be sent or received as a standard or extended ID. For incoming
message requests, we automatically select the format based on the value
of the ID (i.e. greater than 0x7ff is automatically extended).

Fixed #197.