Require an explicit PLATFORM to be set when running Make.
Since the 'make clean' goal was changed to only clean the current
platform, I've had the problem a few times where I'm explicitly setting
the PLATFORM=X when I compile, but not when I clean. It ends up cleaning
the wrong thing, and the firmware isn't rebuilt properly.

This would/will be less of an issue when
https://github.com/openxc/vi-firmware/issues/231 is resolved.