Added support for GET /configure
There was some support previously but it had not been tested.
This commit hacks a NULL onto the received body to satisfy
handleCommand()'s need for a NULL terminator on every JSON frame. This
hack let's us handle a single JSON command in each GET until we modify
the server code.